+++ src/sql/cmd/explore.py
@@ -1,5 +1,4 @@
 from sql.widgets import TableWidget
-from IPython.display import display
 from sql.cmd.cmd_utils import CmdParser
 
 
@@ -18,7 +17,7 @@ def explore(others):
     """
     parser = CmdParser()
     parser.add_argument("-t", "--table", type=str, help="Table name", required=True)
+    parser.add_argument("-s", "--schema", type=str, help="Schema name", required=False)
     args = parser.parse_args(others)
-
-    table_widget = TableWidget(args.table)
-    display(table_widget)
+    table_widget = TableWidget(args.table, args.schema)
+    return table_widget