+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.10.2dev
 
 * [Feature] Improved messages when loading configurations from `pyproject.toml` file.
+* [Feature] Add `--schema/-s` for `%sqlcmd` commands that support `--table/-t` and ensure `--table schema.table` works (#519)
 
 ## 0.10.1 (2023-08-30)